CRNN/requirements.txt
torchvision
fastapi
uvicorn
python-multipart
pillow